Use exact-width integers

+++ include/natalie/string_object.hpp
@@ -127,7 +127,7 @@ public:
     void append(const char *, size_t);
     void append(int);
     void append(long unsigned int);
-    void append(long long int);
+    void append(int64_t);
     void append(long int);
     void append(unsigned int);
     void append(double);
